Streamline messaging: user and IBS will now receive the same output.
No longer user weird overloaded logic of parameters that should be
passed to the IBS, but were also used to construct output-formatting.

E.g. a passed group would become a 'by_group' parameter, but maybe it
should only occur in the comment.

Comments are now always formatted in the actions and only prefixed
in the Request, so the group parameter has only one concern: being
passed to the IBS as 'by_group' or not.